Create TcpEchoServer
